Reworked static data nad page layout
