+++ apps/charterafrica/src/components/Grants/Grant.snap.js
@@ -3,161 +3,245 @@
 exports[`<Grants /> renders unchanged 1`] = `
 <div>
   <div
-    class="MuiBox-root css-0"
+    class="MuiBox-root css-156ejva"
   >
-    <h3
-      class="MuiTypography-root MuiTypography-h3Small css-yo7jji-MuiTypography-root"
-    >
-      Grants
-    </h3>
     <div
-      class="MuiBox-root css-0"
+      class="MuiContainer-root MuiContainer-maxWidthLg MuiContainer-fixed MuiContainer-disableGutters css-u6144g-MuiContainer-root"
     >
+      <h3
+        class="MuiTypography-root MuiTypography-h3Small css-yo7jji-MuiTypography-root"
+      >
+        Fellowships
+      </h3>
       <div
-        class="MuiGrid-root MuiGrid-container css-4cmwq4-MuiGrid-root"
+        class="MuiBox-root css-0"
       >
         <div
-          class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          class="MuiGrid-root MuiGrid-container css-4cmwq4-MuiGrid-root"
         >
-          <h5
-            class="MuiTypography-root MuiTypography-h5SemiBold css-1i8y2l6-MuiTypography-root"
+          <div
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
           >
-            Upcoming Calls
-          </h5>
+            <h5
+              class="MuiTypography-root MuiTypography-h5SemiBold css-1i8y2l6-MuiTypography-root"
+            >
+              open calls
+            </h5>
+          </div>
+          <div
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          >
+            <a
+              class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1ge11hp-MuiTypography-root-MuiLink-root"
+              href="/grants"
+            >
+              View All
+            </a>
+          </div>
         </div>
         <div
-          class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          class="MuiGrid-root MuiGrid-container MuiGrid-spacing-xs-2 css-1j8f29c-MuiGrid-root"
         >
-          <a
-            class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1ge11hp-MuiTypography-root-MuiLink-root"
-            href="/grants"
+          <div
+            class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
           >
-            View All
-          </a>
+            <div
+              class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
+              style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
+            >
+              <img
+                alt="Grant 1"
+                class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
+                height="187.26"
+                src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
+                width="270"
+              />
+              <div
+                class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
+              >
+                <h6
+                  class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
+                >
+                  Democratic Governance in Zambia
+                </h6>
+                <p
+                  class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
+                >
+                  This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
+                </p>
+                <span
+                  class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
+                >
+                  Deadline: 
+                  2023-02-11
+                </span>
+              </div>
+            </div>
+          </div>
         </div>
+        <a
+          class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1qpzgla-MuiTypography-root-MuiLink-root"
+          href="/grants"
+        >
+          View All
+        </a>
       </div>
+      <hr
+        class="MuiDivider-root MuiDivider-fullWidth css-pdvg27-MuiDivider-root"
+      />
       <div
-        class="MuiGrid-root MuiGrid-container MuiGrid-spacing-xs-2 css-1j8f29c-MuiGrid-root"
+        class="MuiBox-root css-0"
       >
         <div
-          class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
+          class="MuiGrid-root MuiGrid-container css-4cmwq4-MuiGrid-root"
         >
           <div
-            class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
-            style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
           >
-            <img
-              alt="Grant 1"
-              class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
-              height="187.26"
-              src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
-              width="270"
-            />
-            <div
-              class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
+            <h5
+              class="MuiTypography-root MuiTypography-h5SemiBold css-1i8y2l6-MuiTypography-root"
             >
-              <h6
-                class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
-              >
-                Democratic Governance in Zambia
-              </h6>
-              <p
-                class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
-              >
-                This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
-              </p>
-              <span
-                class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
-              >
-                Deadline: 
-                2023-02-11
-              </span>
-            </div>
+              closed calls
+            </h5>
+          </div>
+          <div
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          >
+            <a
+              class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1ge11hp-MuiTypography-root-MuiLink-root"
+              href="/grants"
+            >
+              View All
+            </a>
           </div>
         </div>
         <div
-          class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
+          class="MuiGrid-root MuiGrid-container MuiGrid-spacing-xs-2 css-1j8f29c-MuiGrid-root"
         >
           <div
-            class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
-            style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
+            class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
           >
-            <img
-              alt="Grant 2"
-              class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
-              height="187.26"
-              src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
-              width="270"
-            />
             <div
-              class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
+              class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
+              style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
             >
-              <h6
-                class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
+              <img
+                alt="Grant 2"
+                class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
+                height="187.26"
+                src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
+                width="270"
+              />
+              <div
+                class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
               >
-                Democratic Governance in Zambia
-              </h6>
-              <p
-                class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
-              >
-                This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
-              </p>
-              <span
-                class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
-              >
-                Deadline: 
-                2023-02-11
-              </span>
+                <h6
+                  class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
+                >
+                  Democratic Governance in Zambia
+                </h6>
+                <p
+                  class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
+                >
+                  This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
+                </p>
+                <span
+                  class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
+                >
+                  Deadline: 
+                  2023-02-11
+                </span>
+              </div>
             </div>
           </div>
         </div>
-        <div
-          class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
+        <a
+          class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1qpzgla-MuiTypography-root-MuiLink-root"
+          href="/grants"
         >
-          <div
-            class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
-            style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
-          >
-            <img
-              alt="Grant 3"
-              class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
-              height="187.26"
-              src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
-              width="270"
-            />
-            <div
-              class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
-            >
-              <h6
-                class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
-              >
-                Democratic Governance in Zambia
-              </h6>
-              <p
-                class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
-              >
-                This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
-              </p>
-              <span
-                class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
-              >
-                Deadline: 
-                2023-02-11
-              </span>
-            </div>
-          </div>
-        </div>
+          View All
+        </a>
       </div>
-      <a
-        class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1qpzgla-MuiTypography-root-MuiLink-root"
-        href="/grants"
+      <hr
+        class="MuiDivider-root MuiDivider-fullWidth css-pdvg27-MuiDivider-root"
+      />
+      <div
+        class="MuiBox-root css-0"
       >
-        View All
-      </a>
+        <div
+          class="MuiGrid-root MuiGrid-container css-4cmwq4-MuiGrid-root"
+        >
+          <div
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          >
+            <h5
+              class="MuiTypography-root MuiTypography-h5SemiBold css-1i8y2l6-MuiTypography-root"
+            >
+              upcoming calls
+            </h5>
+          </div>
+          <div
+            class="MuiGrid-root MuiGrid-item css-13i4rnv-MuiGrid-root"
+          >
+            <a
+              class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1ge11hp-MuiTypography-root-MuiLink-root"
+              href="/grants"
+            >
+              View All
+            </a>
+          </div>
+        </div>
+        <div
+          class="MuiGrid-root MuiGrid-container MuiGrid-spacing-xs-2 css-1j8f29c-MuiGrid-root"
+        >
+          <div
+            class="MuiGrid-root MuiGrid-item css-muuvl7-MuiGrid-root"
+          >
+            <div
+              class="MuiPaper-root MuiPaper-elevation MuiPaper-rounded MuiPaper-elevation1 MuiCard-root css-uuhpmr-MuiPaper-root-MuiCard-root"
+              style="width: 270px; box-shadow: none; background-color: rgb(254, 250, 237);"
+            >
+              <img
+                alt="Grant 3"
+                class="MuiCardMedia-root MuiCardMedia-media MuiCardMedia-img css-1wcqjo0-MuiCardMedia-root"
+                height="187.26"
+                src="/images/the-charter-project-africa_brandguide_guidedemarque-2-2.jpg"
+                width="270"
+              />
+              <div
+                class="MuiCardContent-root css-1hoqlal-MuiCardContent-root"
+              >
+                <h6
+                  class="MuiTypography-root MuiTypography-h6 css-1p9avxy-MuiTypography-root"
+                >
+                  Democratic Governance in Zambia
+                </h6>
+                <p
+                  class="MuiTypography-root MuiTypography-p1 css-ywxmxe-MuiTypography-root"
+                >
+                  This call will focus on using civic tech solutions to strengthen democratic governance in Zambia.
+                </p>
+                <span
+                  class="MuiTypography-root MuiTypography-caption css-y4met2-MuiTypography-root"
+                >
+                  Deadline: 
+                  2023-02-11
+                </span>
+              </div>
+            </div>
+          </div>
+        </div>
+        <a
+          class="MuiTypography-root MuiTypography-inherit MuiLink-root MuiLink-underlineAlways css-1qpzgla-MuiTypography-root-MuiLink-root"
+          href="/grants"
+        >
+          View All
+        </a>
+      </div>
+      <hr
+        class="MuiDivider-root MuiDivider-fullWidth css-pdvg27-MuiDivider-root"
+      />
     </div>
-    <hr
-      class="MuiDivider-root MuiDivider-fullWidth css-arfqse-MuiDivider-root"
-    />
   </div>
 </div>
 `;
