--- apps/charterafrica/src/components/GrantsFellowships/index.js
@@ -1,3 +1,0 @@
-import GrantsFellowships from "./GrantFellowship";
-
-export default GrantsFellowships;
