+++ apps/charterafrica/src/components/GrantsFellowshipsHeader/GrantsFellowshipsHeader.snap.js
@@ -1,0 +1,22 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<GrantsFellowshipsHeader /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-156ejva"
+  >
+    <div
+      class="MuiContainer-root MuiContainer-maxWidthLg MuiContainer-fixed MuiContainer-disableGutters css-1sbi1ab-MuiContainer-root"
+    >
+      <h3
+        class="MuiTypography-root MuiTypography-h3 css-lj04n5-MuiTypography-root"
+      >
+        Grants and Fellowships
+      </h3>
+      <hr
+        class="MuiDivider-root MuiDivider-fullWidth css-zmwurh-MuiDivider-root"
+      />
+    </div>
+  </div>
+</div>
+`;
