+++ apps/charterafrica/src/components/GrantsFellowshipsHeader/index.js
@@ -1,0 +1,3 @@
+import GrantsFellowshipsHeader from "./GrantFellowshipHeader";
+
+export default GrantsFellowshipsHeader;
